aider: Added a function to cast numbers of type NN to numbers of type II.
# Aider chat conversation:

ASSISTANT: Sure, I can help with that.

To cast a number of type NN to II, we can use the II constructor with the NN number as the first argument and O as the second argument. This is because an integer in this context is represented as the difference of two natural numbers, and casting a natural number to an integer can be thought of as subtracting 0 from the natural number.

Here is the change:

```haskell
Downloads\arithmetic.hs
<<<<<<< HEAD
-- cast numbers of type NN to numbers of type II
ii_nn :: NN -> II
=======
-- cast numbers of type NN to numbers of type II
ii_nn :: NN -> II
ii_nn n = II n O
>>>>>>> updated
```

This function works by taking a natural number n and returning an integer that represents the difference between n and 0.