aider: Cast numbers of type PP to numbers of type II.
# Aider chat conversation:

USER: how about for line 56 to cast numbers of type PP to numbers of type II

ASSISTANT: Sure, I can help with that.

To cast a number of type PP to II, we can use the II constructor with the NN equivalent of the PP number as the first argument and O as the second argument. This is because an integer in this context is represented as the difference of two natural numbers, and casting a positive number to an integer can be thought of as subtracting 0 from the positive number.

Here is the change:

```haskell
Downloads\arithmetic.hs
<<<<<<< HEAD
-- cast numbers of type PP to numbers of type II
ii_pp :: PP -> II
=======
-- cast numbers of type PP to numbers of type II
ii_pp :: PP -> II
ii_pp p = II (nn_pp p) O
>>>>>>> updated
```

This function works by taking a positive number p, converting it to a natural number using the nn_pp function, and then returning an integer that represents the difference between the natural number and 0.