aider: Added function to count the elements on both sides of a chemical equation.
# Aider chat conversation:

USER: now can you help me write a function that will count the elements on both sides of the equation? I ultimately want a function that can balance chemical equations by making sure both sides have the same number of element_symbols on each side

ASSISTANT: Sure, I can help with that. We will need to modify the `custom_blocks.js` file to add a new function that counts the elements on both sides of the equation.

This function will take the chemical equation as input, split it into reactants and products, and then count the number of each element on both sides.

Here's a step-by-step breakdown of what we need to do:

1. Split the equation into reactants and products.
2. For each side of the equation, count the number of each element.
3. Return a dictionary with the counts of each element on both sides.

Here's the code that accomplishes this:

```javascript
OneDrive/Desktop/ChemDSL/custom_blocks.js
<<<<<<< HEAD
=======
function countElements(equation) {
  var sides = equation.split('=>');
  var reactants = sides[0].split('+');
  var products = sides[1].split('+');

  var reactantCounts = {};
  var productCounts = {};

  reactants.forEach(function(reactant) {
    var elements = reactant.match(/[A-Z][a-z]*/g);
    elements.forEach(function(element) {
      if (reactantCounts[element]) {
        reactantCounts[element]++;
      } else {
        reactantCounts[element] = 1;
      }
    });
  });

  products.forEach(function(product) {
    var elements = product.match(/[A-Z][a-z]*/g);
    elements.forEach(function(element) {
      if (productCounts[element]) {
        productCounts[element]++;
      } else {
        productCounts[element] = 1;
      }
    });
  });

  return {
    reactants: reactantCounts,
    products: productCounts
  };
}
>>>>>>> updated
```

This function uses regular expressions to match the element symbols in the equation. It then counts the number of each element on both sides of the equation and returns a dictionary with the counts.

Please note that this function assumes that the equation is formatted correctly and that the element symbols are capitalized correctly. If the equation is not formatted correctly or the element symbols are not capitalized correctly, the function may not work as expected.