aider: Changed the `balance_equation` block to be a value block instead of a statement block.
# Aider chat conversation:

USER: when I try to generate code with the balance_equation block, I get this error: """javascript.js:32
 Uncaught TypeError: Expecting string from statement block: balance_equation
    at JavascriptGenerator$$module$build$src$generators$javascript$javascript_generator.blockToCode (javascript.js:32:14)
    at JavascriptGenerator$$module$build$src$generators$javascript$javascript_generator.workspaceToCode (javascript.js:32:14)
    at generateCode (index.html:46:43)
    at HTMLButtonElement.onclick (index.html:31:42)
blockToCode	@	javascript.js:32
workspaceToCode	@	javascript.js:32
generateCode	@	index.html:46
onclick	@	index.html:31"""

ASSISTANT: The error message indicates that the `balance_equation` block is expected to return a string, but it's not. This is because the `balance_equation` block is defined as a statement block (it doesn't return a value), but it's being used as a value block (it returns a value).

To fix this, we need to change the `balance_equation` block to be a value block. This involves changing the `setPreviousStatement` method to `setOutput` method in the `init` function of the `balance_equation` block.

Here is the *edit block* for the changes:

```javascript
OneDrive/Desktop/ChemDSL/custom_blocks.js
<<<<<<< HEAD
Blockly.Blocks['balance_equation'] = {
	init: function() {
		this.appendValueInput("BALANCEDEQUATION")
			.setCheck("EquationOutput")
			.appendField("EQ Balance");
		this.setInputsInline(true);
		this.setPreviousStatement(true, "ChemFunction");
		this.setColour(230);
		this.setTooltip("Insert a chemical equation to check for balance");
		this.setHelpUrl("");
	}
};
=======
Blockly.Blocks['balance_equation'] = {
	init: function() {
		this.appendValueInput("BALANCEDEQUATION")
			.setCheck("EquationOutput")
			.appendField("EQ Balance");
		this.setInputsInline(true);
		this.setOutput(true, "ChemFunction");
		this.setColour(230);
		this.setTooltip("Insert a chemical equation to check for balance");
		this.setHelpUrl("");
	}
};
>>>>>>> updated
```

Please confirm these changes.