aider: Changed `balance_equation` block to a value block while keeping its shape and allowing it to be attached to the block above it.
# Aider chat conversation:

USER: I don't want to change the shape of the block though so I need the setPreviousStatement to stay

ASSISTANT: Understood. The `setPreviousStatement` method is used to allow this block to be attached to the block above it. If you want to keep the shape of the block and allow it to be attached to the block above it, we should keep this method call.

However, if you want the `balance_equation` block to return a value (i.e., make it a value block), we need to add the `setOutput` method call. This method allows the block to be attached to the input of another block.

Here is the change that needs to be made to the `balance_equation` block in the `custom_blocks.js` file:

```javascript
OneDrive/Desktop/ChemDSL/custom_blocks.js
<<<<<<< HEAD
Blockly.Blocks['balance_equation'] = {
	init: function() {
		this.appendValueInput("BALANCEDEQUATION")
			.setCheck("EquationOutput")
			.appendField("EQ Balance");
		this.setInputsInline(true);
		this.setOutput(true, "ChemFunction");
		this.setColour(230);
		this.setTooltip("Insert a chemical equation to check for balance");
		this.setHelpUrl("");
	}
};
=======
Blockly.Blocks['balance_equation'] = {
	init: function() {
		this.appendValueInput("BALANCEDEQUATION")
			.setCheck("EquationOutput")
			.appendField("EQ Balance");
		this.setInputsInline(true);
		this.setOutput(true, "ChemFunction");
		this.setPreviousStatement(true, null);
		this.setColour(230);
		this.setTooltip("Insert a chemical equation to check for balance");
		this.setHelpUrl("");
	}
};
>>>>>>> updated
```

This change will make the `balance_equation` block a value block while keeping its shape and allowing it to be attached to the block above it.