aider: Modified the `generateCode` function to check the type of the top block in the Blockly workspace and call the appropriate function based on the block type.
# Aider chat conversation:

USER: but won't that call the balance_equation block every time I press generateCode? Even when I'm not trying to balance an equation? That doesn't seem scalable.

ASSISTANT: You're correct. The current implementation will call the `balanceEquation` function every time the `generateCode` function is called, regardless of the type of block used. This is not scalable and could lead to errors if the `balanceEquation` function is called with an inappropriate argument.

A better approach would be to modify the `generateCode` function to check the type of the top block in the Blockly workspace and call the appropriate function based on the block type.

Here are the changes:

```html
OneDrive/Desktop/ChemDSL/index.html
<<<<<<< HEAD
        function generateCode() {
            var code = Blockly.JavaScript.workspaceToCode(workspace);
            var result = balanceEquation(code);
            document.getElementById('codeDiv').innerText = result;
        }
=======
        function generateCode() {
            var topBlocks = workspace.getTopBlocks();
            var code = Blockly.JavaScript.workspaceToCode(workspace);
            var result;
            if (topBlocks.length > 0 && topBlocks[0].type === 'balance_equation') {
                result = balanceEquation(code);
            } else {
                result = code;
            }
            document.getElementById('codeDiv').innerText = result;
        }
>>>>>>> updated
```

This change will check the type of the top block in the Blockly workspace. If the top block is a `balance_equation` block, it will call the `balanceEquation` function with the generated code as the argument. Otherwise, it will simply display the generated code. This ensures that the `balanceEquation` function is only called when appropriate.